Move logging to main
